# Optimization with Python comprehensions