% `ruff --select=C4,PERF,UP027,UP028 --statistics ` # https://docs.astral.sh/ruff
```
19	C419   	[*] Unnecessary list comprehension
 7	UP028  	[*] Replace `yield` over `for` loop with `yield from`
 3	C403   	[*] Unnecessary `list` comprehension (rewrite as a `set` comprehension)
 3	C405   	[*] Unnecessary `list` literal (rewrite as a `set` literal)
 3	C414   	[*] Unnecessary `list` call within `set()`
 3	PERF401	[ ] Use a list comprehension to create a transformed list
 3	UP027  	[*] Replace unpacked list comprehension with a generator expression
 2	C417   	[*] Unnecessary `map` usage (rewrite using a `list` comprehension)
 2	PERF203	[ ] `try`-`except` within a loop incurs performance overhead
 1	C410   	[*] Unnecessary `list` literal passed to `list()` (remove the outer call to `list()`)
 1	C416   	[*] Unnecessary `list` comprehension (rewrite using `list()`)
 1	PERF102	[*] When using only the values of a dict use the `values()` method
```
% `ruff --select=C4,PERF,UP027,UP028 --fix --unsafe-fixes ` # Plus manual fixes for PERF401

% `ruff --select=C4,PERF,UP027,UP028 --statistics`
```
2	PERF203	`try`-`except` within a loop incurs performance overhead
```
% `ruff rule C419`
# unnecessary-comprehension-any-all (C419)

Derived from the **flake8-comprehensions** linter.

Fix is sometimes available.

## What it does
Checks for unnecessary list comprehensions passed to `any` and `all`.

## Why is this bad?
`any` and `all` take any iterators, including generators. Converting a generator to a list
by way of a list comprehension is unnecessary and reduces performance due to the
overhead of creating the list.

For example, compare the performance of `all` with a list comprehension against that
of a generator (~40x faster here):

```console
In [1]: %timeit all([i for i in range(1000)])
8.14 µs ± 25.4 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [2]: %timeit all(i for i in range(1000))
212 ns ± 0.892 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

## Examples
```python
any([x.id for x in bar])
all([x.id for x in bar])
```

Use instead:
```python
any(x.id for x in bar)
all(x.id for x in bar)
```

## Fix safety
This rule's fix is marked as unsafe, as it may occasionally drop comments
when rewriting the comprehension. In most cases, though, comments will be
preserved.